feat: add ExpenseRefusalPopupComponent
